refactor: update import paths to use absolute paths and improve code readability
