+++ package.json
@@ -14,6 +14,7 @@
     "seed:admin": "tsx app/.server/seeds/seed-admin-user.ts"
   },
   "dependencies": {
+    "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "@react-router/node": "^7.5.3",
     "@react-router/serve": "^7.5.3",
@@ -34,12 +35,12 @@
     "@types/node": "^20",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
+    "dotenv": "^17.0.1",
     "tailwindcss": "^4.1.4",
     "tsx": "^4.20.3",
     "tw-animate-css": "^1.3.5",
     "typescript": "^5.8.3",
     "vite": "^6.3.3",
-    "vite-tsconfig-paths": "^5.1.4",
-    "dotenv": "^17.0.1"
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
